+++ source/ScrollableContainerNotReadyError.js
@@ -1,0 +1,5 @@
+export default class ScrollableContainerNotReadyError extends Error {
+	constructor() {
+		super('[virtual-scroller] Scrollable container not found');
+	}
+}